+++ src/commands/index.ts
@@ -1,3 +1,5 @@
+export * from './Excerpts';
+export * from './Freeform';
 export * from './Headline';
 export * from './Metadata';
 export * from './RevisionText';