+++ src/extension.ts
@@ -1,7 +1,8 @@
+import { Freeform } from './commands/Freeform';
 import { Metadata } from './commands/Metadata';
 import { RevisionText } from './commands/RevisionText';
 import * as vscode from 'vscode';
-import { Headline, SeoOptimization, TextEdits } from './commands';
+import { Excerpts, Headline, SeoOptimization, TextEdits } from './commands';
 import { SparkupContentProvider } from './provider';
 
 
@@ -10,10 +11,25 @@ export function activate(context: vscode.ExtensionContext) {
 
 	new SparkupContentProvider(context).init();
 
+	/**
+	 * Headlines
+	 */
 	context.subscriptions.push(
 		vscode.commands.registerCommand('vscode-sparkup.headline', Headline.generate)
 	);
 
+
+	/**
+	 * Excerpts
+	 */
+	context.subscriptions.push(
+		vscode.commands.registerCommand('vscode-sparkup.excerpts', Excerpts.process)
+	);
+
+
+	/**
+	 * Text Edits
+	 */
 	context.subscriptions.push(
 		vscode.commands.registerCommand('vscode-sparkup.fixSpelling', TextEdits.fixSpelling)
 	);
@@ -26,18 +42,34 @@ export function activate(context: vscode.ExtensionContext) {
 		vscode.commands.registerCommand('vscode-sparkup.simplify', TextEdits.simplify)
 	);
 
+
+	/**
+	 * Freeform
+	 */
 	context.subscriptions.push(
-		vscode.commands.registerCommand('vscode-sparkup.freeform', TextEdits.freeform)
+		vscode.commands.registerCommand('vscode-sparkup.freeform', Freeform.freeform)
 	);
 
-	context.subscriptions.push(
-		vscode.commands.registerCommand('vscode-sparkup.seoOptimize', SeoOptimization.optimize)
-	);
 
+	/**
+	 * Revision Text
+	 */
 	context.subscriptions.push(
 		vscode.commands.registerCommand('vscode-sparkup.revisionText', RevisionText.revision)
 	);
 
+
+	/**
+	 * SEO Optimization
+	 */
+	context.subscriptions.push(
+		vscode.commands.registerCommand('vscode-sparkup.seoOptimize', SeoOptimization.optimize)
+	);
+
+
+	/**
+	 * Categorization
+	 */
 	context.subscriptions.push(
 		vscode.commands.registerCommand('vscode-sparkup.categorizeAndTag', Metadata.process)
 	);