+++ src/models/FreeformInstruction.ts
@@ -1,0 +1,5 @@
+export interface FreeformInstruction {
+  title: string;
+  instruction: string;
+  nrOfResults: number;
+}